Fix overflow error in interpolation search, finish cleanup

+++ src/main/java/net/adeptropolis/nephila/graph/implementations/primitives/search/InterpolationSearch.java
@@ -49,7 +49,7 @@ public class InterpolationSearch {
     int midVal;
 
     while ((highVal != lowVal) && (key >= lowVal) && (key <= highVal)) {
-      mid = low + ((key - lowVal) * (high - low) / (highVal - lowVal));
+      mid = (int)(low + (((key - lowVal) * (long)(high - low)) / (highVal - lowVal)));
       midVal = array[mid];
       if (key > midVal) {
         low = mid + 1;
